[WIP] day 17
Kind of gave up on day 16, but day 17 is very interesting. On the
subreddit, I saw someone writing up explanations on how they solved each
problem on their website. I find it to be an interesting exercise, so
I'm trying it for this problem. I'll try to do it at the same time as I
solve the problem for the next ones.